pyproject.toml
[project]
name = "chaos-api"
version = "0.1.0"
description = "A chaos API that generates fake REST responses using LLM"
authors = [
    {name = "John Sosoka"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "langchain (>=1.2.13,<2.0.0)",
    "langchain-openai (>=1.1.11,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)"
]

[project.optional-dependencies]
dev = [
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-asyncio (>=0.23.0,<0.24.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "httpx (>=0.27.0,<0.28.0)",
    "ruff (>=0.5.0,<0.6.0)",
    "mypy (>=1.10.0,<2.0.0)",
    "pre-commit (>=3.7.0,<4.0.0)",
    "types-python-dateutil"
]

[tool.poetry]
packages = [{include = "chaos_api", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = ["-v"]

# ruff configuration
[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["SIM117"]  # Allow nested with statements in tests

# mypy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = ["langchain.*", "langchain_openai.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["chaos_api.config"]
ignore_errors = true
